startup - apply default_cmdline before jsk_used set_bit, rw first run message
This fixes an issue where if --enablejsk is applied through default_cmdline before jsk_used is set, jsk_used will not be set.
This was caused by --enablejsk only being appended to sys.argv following the check for --enablejsk in sys.argv.
This check would always return False if --enablejsk was in default_cmdline, causing jsk_used to not be written to config.

This also rewrites a substantial portion of the 2.0 first run message to reflect the current project state.