begin logging impl cleanup, see desc
This commit improves the logging implementation for some core files.
base.maximilian.logger is no longer referenced in these files but is still set for now.
Deprecated methods are no longer used, and loggers are not passed as parameters anymore.
Instead, logging.getLogger is called every time a logger is needed within a function's scope.
For things that use the root logger, getLogger is passed constants.GlobalConstants.ROOT_LOGGER_NAME to access the root logger instance.
See https://docs.python.org/3/howto/logging-cookbook.html#using-loggers-as-attributes-in-a-class-or-passing-them-as-parameters